update to latest mupdf

+++ mupdf/platform/gl/gl-font.c
@@ -61,8 +61,8 @@ static fz_font *g_font = NULL;
 static void clear_font_cache(void)
 {
 #if PADDING > 0
-	unsigned char *zero = malloc(g_cache_w * g_cache_h);
-	memset(zero, 0, g_cache_w * g_cache_h);
+	unsigned char *zero = malloc((size_t)g_cache_w * g_cache_h);
+	memset(zero, 0, (size_t)g_cache_w * g_cache_h);
 	glBindTexture(GL_TEXTURE_2D, g_cache_tex);
 	glTexSubImage2D(GL_TEXTURE_2D, 0, 0, 0, g_cache_w, g_cache_h, GL_ALPHA, GL_UNSIGNED_BYTE, zero);
 	free(zero);
