+++ mupdf/source/fitz/load-gif.c
@@ -275,7 +275,7 @@ gif_read_tbid(fz_context *ctx, struct info *info, const unsigned char *p, const 
 		lzwstm = fz_open_lzwd(ctx, stm, 0, mincodesize + 1, 1, 1);
 
 		uncompressed = fz_read_all(ctx, lzwstm, 0);
-		if (uncompressed->len < info->image_width * info->image_height)
+		if (uncompressed->len < (size_t)info->image_width * info->image_height)
 			fz_throw(ctx, FZ_ERROR_GENERIC, "premature end in compressed table based image data in gif image");
 
 		if (info->has_lct)
@@ -474,7 +474,7 @@ gif_read_image(fz_context *ctx, struct info *info, const unsigned char *p, size_
 
 	fz_try(ctx)
 	{
-		info->mask = fz_calloc(ctx, info->width * info->height, 1);
+		info->mask = fz_calloc(ctx, (size_t)info->width * info->height, 1);
 
 		/* Read optional global color table */
 		if (info->has_gct)
@@ -485,7 +485,7 @@ gif_read_image(fz_context *ctx, struct info *info, const unsigned char *p, size_
 			p = gif_read_gct(ctx, info, p, end);
 			bp = &info->gct[info->gct_background * 3];
 
-			memset(info->mask, 0x01, info->width * info->height);
+			memset(info->mask, 0x01, (size_t)info->width * info->height);
 
 			for (y = 0; y < info->height; y++)
 				for (x = 0; x < info->width; x++, dp += 4)
