+++ mupdf/source/fitz/output-pclm.c
@@ -120,8 +120,8 @@ pclm_write_header(fz_context *ctx, fz_band_writer *writer_, fz_colorspace *cs)
 
 	fz_free(ctx, writer->stripbuf);
 	fz_free(ctx, writer->compbuf);
-	writer->stripbuf = Memento_label(fz_malloc(ctx, w * sh * n), "pclm_stripbuf");
-	writer->complen = fz_deflate_bound(ctx, w * sh * n);
+	writer->stripbuf = Memento_label(fz_malloc(ctx, (size_t)w * sh * n), "pclm_stripbuf");
+	writer->complen = fz_deflate_bound(ctx, (size_t)w * sh * n);
 	writer->compbuf = Memento_label(fz_malloc(ctx, writer->complen), "pclm_compbuf");
 
 	/* Send the file header on the first page */
@@ -186,7 +186,7 @@ flush_strip(fz_context *ctx, pclm_band_writer *writer, int fill)
 	fz_output *out = writer->super.out;
 	int w = writer->super.w;
 	int n = writer->super.n;
-	size_t len = w*n*fill;
+	size_t len = (size_t)w*n*fill;
 
 	/* Buffer is full, compress it and write it. */
 	if (writer->options.compress)
@@ -220,7 +220,9 @@ pclm_write_band(fz_context *ctx, fz_band_writer *writer_, int stride, int band_s
 	for (line = 0; line < band_height; line++)
 	{
 		int dstline = (band_start+line) % sh;
-		memcpy(writer->stripbuf + w*n*dstline, sp + line * w*n, w*n);
+		memcpy(writer->stripbuf + (size_t)w*n*dstline,
+			   sp + (size_t)line * w * n,
+			   (size_t)w * n);
 		if (dstline+1 == sh)
 			flush_strip(ctx, writer, dstline+1);
 	}
