+++ mupdf/source/tools/muraster.c
@@ -977,7 +977,7 @@ initialise_banding(fz_context *ctx, render_details *render, int color)
 	}
 
 	w = render->ibounds.x1 - render->ibounds.x0;
-	min_band_mem = bpp * w * min_band_height;
+	min_band_mem = (size_t)bpp * w * min_band_height;
 	reps = (int)(max_band_memory / min_band_mem);
 	if (reps < 1)
 		reps = 1;
